treewide: Revert all strlcpy usage fixes
A breakage was discovered in diag driver with this kind of 'fix' in place.
This causes userspace that depend on diag to stop working with this error
spam in dmesg:

  [53203.332480] diag: Unable to send msg masks to peripheral 0, error = -11

Proper fixes for diag driver are present in msm-4.14, while for MMC can
be reverted back to pre-CAF state with 3 patches from bluecross.

For the rest: Correct the size argument instead.

* Revert "soc/qcom: msm_minidump: Fix strlcpy usage"
  This reverts commit e9517ca534b766e6159e10050b8c3fd24472a828.
* Revert "mmc: sdio: Fix strlcpy usage"
  This reverts commit c57b11db3c2910387eed8ea6f28e84f196ac331b.
* Revert "soc: qcom: glink_debugfs: Fix strlcpy usage"
  This reverts commit 4a14100a022855eb8eb7b4a2161dac032a3283ef.
* Revert "char: diag: Fix strlcpy usage"
  This reverts commit d518c2609c002e55d926c425b059033debc4b94a.

Signed-off-by: Albert I <kras@raphielgang.org>